+++ CHANGES.md
@@ -1,7 +1,8 @@
 ## 1.1.3 (unreleased)
 
 
-- Nothing changed yet.
+- Check if realm exists and environment variables are set before migration
+  [remdub]
 
 
 ## 1.1.2 (2025-08-27)